requirements.txt
asgiref==3.2.3
dataclasses==0.6
dj-database-url==0.5.0
Django==3.0.3
django-cors-headers==3.2.1
django-filter==2.2.0
django-heroku==0.3.1
django-ordered-model==3.4.1
gunicorn==20.0.4
psycopg2==2.8.4
pycodestyle==2.6.0
pyflakes==2.2.0
pylint==2.4.4
pylint-django==2.0.13
pylint-plugin-utils==0.6
pyparsing==2.4.7
python-dateutil==2.8.1
requests==2.22.0
requests-oauthlib==1.3.0
sqlparse==0.2.4
urllib3==1.25.7
virtualenv==16.7.9
whitenoise==5.0.1
